chore: add core package tests

+++ tests/utils/MockFsPackage/MockPackage.ts
@@ -1,4 +1,5 @@
-import { ErrorResult, Package, Part } from "../../src";
+/* eslint-disable security/detect-object-injection */
+import { ErrorResult, Package, Part } from "../../../src";
 import { Volume } from "memfs";
 import { Err, Ok, Result } from "ts-results";
 // eslint-disable-next-line ava/no-import-test-files
@@ -8,35 +9,42 @@ export class MockPackage extends Package {
 	filesystem;
 	mimes: Record<string, string> = {};
 
+	configuredPartErrors: Record<
+		string,
+		{
+			e: ErrorResult;
+			operation: ("DELETE" | "EXIST" | "GET")[];
+		}
+	> = {};
+
 	constructor() {
 		super();
 		this.filesystem = Volume.fromJSON({});
 	}
 
 	protected deletePartCore(name: string): Promise<Result<void, ErrorResult>> {
-		try {
-			// eslint-disable-next-line security/detect-object-injection
-			delete this.mimes[name];
-			this.filesystem.unlinkSync(name);
-		} catch (e) {
-			return Promise.resolve(
-				Err({
-					Category: "PART",
-					Code: "INTERNAL",
-					Error: e
-				} as ErrorResult)
-			);
+		if (this.configuredPartErrors[name]?.operation.includes("DELETE")) {
+			return Promise.resolve(Err(this.configuredPartErrors[name].e));
 		}
+
+		delete this.mimes[name];
+		this.filesystem.unlinkSync(name);
 
 		return Promise.resolve(Ok.EMPTY);
 	}
 
 	protected getPartCore(name: string): Promise<Result<Part, ErrorResult>> {
-		// eslint-disable-next-line security/detect-object-injection
+		if (this.configuredPartErrors[name]?.operation.includes("GET")) {
+			return Promise.resolve(Err(this.configuredPartErrors[name].e));
+		}
+
 		return Promise.resolve(Ok<Part>(new MockPart(this, name, this.mimes[name])));
 	}
 
 	protected partExistsCore(name: string): Promise<Result<boolean, ErrorResult>> {
+		if (this.configuredPartErrors[name]?.operation.includes("EXIST")) {
+			return Promise.resolve(Err(this.configuredPartErrors[name].e));
+		}
 		return Promise.resolve(Ok(this.filesystem.existsSync(name)));
 	}
 }
